+++ src/Interface/TextList.h
@@ -43,7 +43,7 @@ class TextList : public InteractiveSurface
 {
 private:
 	std::vector< std::vector<Text*> > _texts;
-	std::vector<int> _columns, _rows;
+	std::vector<size_t> _columns, _rows;
 	Font *_big, *_small, *_font;
 	Language *_lang;
 	size_t _scroll, _visibleRows, _selRow;
@@ -82,17 +82,17 @@ public:
 	/// Unpresses the surface.
 	void unpress(State *state);
 	/// Sets the text color of a certain cell.
-	void setCellColor(int row, int column, Uint8 color);
+	void setCellColor(size_t row, size_t column, Uint8 color);
 	/// Sets the text color of a certain row.
-	void setRowColor(int row, Uint8 color);
+	void setRowColor(size_t row, Uint8 color);
 	/// Gets the text of a certain cell.
-	std::wstring getCellText(int row, int column) const;
+	std::wstring getCellText(size_t row, size_t column) const;
 	/// Sets the text of a certain cell.
-	void setCellText(int row, int column, const std::wstring &text);
+	void setCellText(size_t row, size_t column, const std::wstring &text);
 	/// Gets the X position of a certain column.
-	int getColumnX(int column) const;
+	int getColumnX(size_t column) const;
 	/// Gets the Y position of a certain row.
-	int getRowY(int row) const;
+	int getRowY(size_t row) const;
 	/// Gets the amount of text in the list.
 	size_t getTexts() const;
 	/// Gets the amount of rows in the list.
@@ -184,7 +184,7 @@ public:
 	/// Special handling for mouse hovering out.
 	void mouseOut(Action *action, State *state);
 	/// get the scroll depth
-	int getScroll();
+	size_t getScroll();
 	/// set the scroll depth
 	void scrollTo(size_t scroll);
 	/// Attaches this button to a combobox.
